Added Formating

+++ test/CradleRaise.t.sol
@@ -4,12 +4,18 @@ pragma solidity ^0.8.20;
 import {Test, console2} from "forge-std/Test.sol";
 import {CradleFactory} from "../src/CradleFactory.sol";
 // Import errors from CradleRaise
-import { 
+import {
     CradleRaise,
     // List imported errors used in tests:
-    InvalidMerkleProof, BelowMinAllocation, ExceedsMaxAllocation, SaleNotActive,
-    ExceedsHardCap, SaleIsFinalizedOrCancelled, CancellationWindowPassed, 
-    SaleNotEnded, SaleNotFinalized
+    InvalidMerkleProof,
+    BelowMinAllocation,
+    ExceedsMaxAllocation,
+    SaleNotActive,
+    ExceedsHardCap,
+    SaleIsFinalizedOrCancelled,
+    CancellationWindowPassed,
+    SaleNotEnded,
+    SaleNotFinalized
 } from "../src/CradleRaise.sol";
 import {IERC20} from "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 import {SafeERC20} from "@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol";
@@ -37,17 +43,17 @@ contract CradleRaiseTest is Test {
     address internal constant user3 = address(0x3333333333333333333333333333333333333333);
 
     // Raise Parameters (adjust as needed for tests)
-    uint8   internal constant TOKEN_DECIMALS = 18;
-    uint8   internal constant ACCEPTED_TOKEN_DECIMALS = 6; // e.g., USDC
-    uint256 internal pricePerWholeToken = 1 * (10**ACCEPTED_TOKEN_DECIMALS); // Price: 1 acceptedToken per 1 whole token
+    uint8 internal constant TOKEN_DECIMALS = 18;
+    uint8 internal constant ACCEPTED_TOKEN_DECIMALS = 6; // e.g., USDC
+    uint256 internal pricePerWholeToken = 1 * (10 ** ACCEPTED_TOKEN_DECIMALS); // Price: 1 acceptedToken per 1 whole token
     uint256 internal presaleStart;
     uint256 internal publicSaleStart;
     uint256 internal endTime;
     bytes32 internal merkleRoot = bytes32(0); // Default to no presale
-    uint16  internal feeBps = 500; // 5%
-    uint256 internal maxRaiseAccepted = 100_000 * (10**ACCEPTED_TOKEN_DECIMALS); // 100k acceptedToken
-    uint256 internal minAllocToken = 100 * (10**TOKEN_DECIMALS); // 100 units of token
-    uint256 internal maxAllocToken = 1000 * (10**TOKEN_DECIMALS); // 1000 units of token
+    uint16 internal feeBps = 500; // 5%
+    uint256 internal maxRaiseAccepted = 100_000 * (10 ** ACCEPTED_TOKEN_DECIMALS); // 100k acceptedToken
+    uint256 internal minAllocToken = 100 * (10 ** TOKEN_DECIMALS); // 100 units of token
+    uint256 internal maxAllocToken = 1000 * (10 ** TOKEN_DECIMALS); // 1000 units of token
 
     // Correct root derived from manual check in test: 0x4beda981c9d34f2dd099131be6049a1d87676d227e63f4a409ee629043314b4f
     bytes32 constant TEST_MERKLE_ROOT = 0x4beda981c9d34f2dd099131be6049a1d87676d227e63f4a409ee629043314b4f;
@@ -65,11 +71,19 @@ contract CradleRaiseTest is Test {
 
         vm.prank(factoryOwner);
         address specificRaiseAddress = factory.createRaise(
-            address(token), address(acceptedToken), pricePerWholeToken,
-            _presaleStart, _publicSaleStart, _endTime, // Future times
+            address(token),
+            address(acceptedToken),
+            pricePerWholeToken,
+            _presaleStart,
+            _publicSaleStart,
+            _endTime, // Future times
             TEST_MERKLE_ROOT, // Use the pre-calculated root
-            projectWallet, feeRecipient, feeBps,
-            maxRaiseAccepted, minAllocToken, maxAllocToken
+            projectWallet,
+            feeRecipient,
+            feeBps,
+            maxRaiseAccepted,
+            minAllocToken,
+            maxAllocToken
         );
         CradleRaise presaleRaise = CradleRaise(payable(specificRaiseAddress));
 
@@ -92,7 +106,7 @@ contract CradleRaiseTest is Test {
         vm.warp(presaleRaise.presaleStart() + 1 seconds);
 
         uint256 amountToBuy = minAllocToken;
-        uint256 expectedAccepted = (amountToBuy * pricePerWholeToken) / (10**TOKEN_DECIMALS);
+        uint256 expectedAccepted = (amountToBuy * pricePerWholeToken) / (10 ** TOKEN_DECIMALS);
 
         uint256 initialUserBalance = acceptedToken.balanceOf(user1);
         uint256 initialRaiseBalance = acceptedToken.balanceOf(address(presaleRaise));
@@ -101,8 +115,16 @@ contract CradleRaiseTest is Test {
         bytes32 leafUser1 = keccak256(abi.encodePacked(user1));
         bytes32 leafUser2 = keccak256(abi.encodePacked(user2));
         // Check against expected values just in case
-        assertEq(leafUser1, bytes32(0xe2c07404b8c1df4c46226425cac68c28d27a766bbddce62309f36724839b22c0), "Leaf User 1 mismatch");
-        assertEq(leafUser2, bytes32(0x2ab0a4443bbea3fbe4d0e1503d11ff1367842fb0c8b28a5c8550f27599a40751), "Leaf User 2 mismatch");
+        assertEq(
+            leafUser1,
+            bytes32(0xe2c07404b8c1df4c46226425cac68c28d27a766bbddce62309f36724839b22c0),
+            "Leaf User 1 mismatch"
+        );
+        assertEq(
+            leafUser2,
+            bytes32(0x2ab0a4443bbea3fbe4d0e1503d11ff1367842fb0c8b28a5c8550f27599a40751),
+            "Leaf User 2 mismatch"
+        );
 
         // Construct proof for user1 dynamically
         bytes32[] memory dynamicProofUser1 = new bytes32[](1);
@@ -122,13 +144,14 @@ contract CradleRaiseTest is Test {
         bytes32 expectedTopic0 = keccak256("Contributed(address,uint256,uint256)");
         bytes32 expectedTopic1 = bytes32(uint256(uint160(user1))); // Address padded
 
-        for (uint i = 0; i < entries.length; i++) {
-            if (entries[i].topics.length > 1 && // Ensure topic1 exists
-                entries[i].topics[0] == expectedTopic0 && 
-                entries[i].topics[1] == expectedTopic1 && // Check contributor topic
-                entries[i].emitter == address(presaleRaise)) 
-            {
-                (uint256 emittedTokenAmount, uint256 emittedAcceptedAmount) = abi.decode(entries[i].data, (uint256, uint256));
+        for (uint256 i = 0; i < entries.length; i++) {
+            if (
+                entries[i].topics.length > 1 // Ensure topic1 exists
+                    && entries[i].topics[0] == expectedTopic0 && entries[i].topics[1] == expectedTopic1 // Check contributor topic
+                    && entries[i].emitter == address(presaleRaise)
+            ) {
+                (uint256 emittedTokenAmount, uint256 emittedAcceptedAmount) =
+                    abi.decode(entries[i].data, (uint256, uint256));
                 assertEq(emittedTokenAmount, amountToBuy, "Log Data: Token Amount Mismatch");
                 assertEq(emittedAcceptedAmount, expectedAccepted, "Log Data: Accepted Amount Mismatch");
                 break;
@@ -137,10 +160,16 @@ contract CradleRaiseTest is Test {
                 fail(); // Fail unconditionally if event not found
             }
         }
-        
+
         // Assert State Changes
-        assertEq(acceptedToken.balanceOf(user1), initialUserBalance - expectedAccepted, "Presale: User balance incorrect");
-        assertEq(acceptedToken.balanceOf(address(presaleRaise)), initialRaiseBalance + expectedAccepted, "Presale: Raise balance incorrect");
+        assertEq(
+            acceptedToken.balanceOf(user1), initialUserBalance - expectedAccepted, "Presale: User balance incorrect"
+        );
+        assertEq(
+            acceptedToken.balanceOf(address(presaleRaise)),
+            initialRaiseBalance + expectedAccepted,
+            "Presale: Raise balance incorrect"
+        );
         assertEq(presaleRaise.contributions(user1), amountToBuy, "Presale: User contribution incorrect");
     }
 
@@ -183,7 +212,7 @@ contract CradleRaiseTest is Test {
 
         // Act
         vm.prank(nonWhitelistedUser);
-        presaleRaise.contribute(amountToBuy, new bytes32[](0)); 
+        presaleRaise.contribute(amountToBuy, new bytes32[](0));
     }
 
     // TODO: Add more presale tests...
@@ -192,7 +221,7 @@ contract CradleRaiseTest is Test {
     // Test Setup
     function setUp() public virtual {
         console2.log("--- Starting setUp ---");
-        
+
         // Ensure block.timestamp is reasonable before adding to it
         vm.warp(10 days); // Initial timestamp = 864000
         console2.log("Warped block.timestamp to:", block.timestamp);
@@ -239,7 +268,7 @@ contract CradleRaiseTest is Test {
 
         // 5. Mint AcceptedToken to Users
         console2.log("Minting accepted tokens to users...");
-        uint256 userInitialBalance = 5000 * (10**ACCEPTED_TOKEN_DECIMALS); // Give users 5k acceptedToken
+        uint256 userInitialBalance = 5000 * (10 ** ACCEPTED_TOKEN_DECIMALS); // Give users 5k acceptedToken
         acceptedToken.mint(user1, userInitialBalance);
         acceptedToken.mint(user2, userInitialBalance);
         acceptedToken.mint(user3, userInitialBalance);
@@ -303,9 +332,9 @@ contract CradleRaiseTest is Test {
     function test_PublicContribute_Success() public {
         // Arrange
         vm.warp(publicSaleStart);
-        uint256 tokenAmountToBuy = 150 * (10**TOKEN_DECIMALS); // Between min (100) and max (1000)
+        uint256 tokenAmountToBuy = 150 * (10 ** TOKEN_DECIMALS); // Between min (100) and max (1000)
         // Calculate expected accepted token amount based on contract's logic
-        uint256 expectedAcceptedToken = (tokenAmountToBuy * pricePerWholeToken) / (10**TOKEN_DECIMALS);
+        uint256 expectedAcceptedToken = (tokenAmountToBuy * pricePerWholeToken) / (10 ** TOKEN_DECIMALS);
 
         uint256 initialUserBalance = acceptedToken.balanceOf(user1);
         uint256 initialRaiseBalance = acceptedToken.balanceOf(address(raise));
@@ -324,7 +353,11 @@ contract CradleRaiseTest is Test {
         // Assert
         // Balances
         assertEq(acceptedToken.balanceOf(user1), initialUserBalance - expectedAcceptedToken, "User balance incorrect");
-        assertEq(acceptedToken.balanceOf(address(raise)), initialRaiseBalance + expectedAcceptedToken, "Raise balance incorrect");
+        assertEq(
+            acceptedToken.balanceOf(address(raise)),
+            initialRaiseBalance + expectedAcceptedToken,
+            "Raise balance incorrect"
+        );
 
         // Contract State
         assertEq(raise.totalAcceptedTokenRaised(), initialTotalRaised + expectedAcceptedToken, "Total raised incorrect");
@@ -423,7 +456,7 @@ contract CradleRaiseTest is Test {
 
     function test_Fail_Contribute_ExceedsHardCap() public {
         // Arrange: Deploy a new raise with a specific low hard cap for this test
-        uint256 specificMaxRaise = 150 * (10**ACCEPTED_TOKEN_DECIMALS); // 150 USDC
+        uint256 specificMaxRaise = 150 * (10 ** ACCEPTED_TOKEN_DECIMALS); // 150 USDC
         // Use other standard parameters from setUp
         vm.prank(factoryOwner);
         address specificRaiseAddress = factory.createRaise(
@@ -455,11 +488,18 @@ contract CradleRaiseTest is Test {
         // minAllocToken = 100 * 10**18
         vm.prank(user1);
         specificRaise.contribute(minAllocToken, new bytes32[](0));
-        assertEq(specificRaise.totalAcceptedTokenRaised(), 100 * (10**ACCEPTED_TOKEN_DECIMALS));
+        assertEq(specificRaise.totalAcceptedTokenRaised(), 100 * (10 ** ACCEPTED_TOKEN_DECIMALS));
 
         // Act & Assert: User 2 attempts to contribute 100 TKN (requires 100 USDC)
         // Total would be 200 USDC, exceeding the 150 USDC cap.
-        vm.expectRevert(abi.encodeWithSelector(ExceedsHardCap.selector, 100 * (10**ACCEPTED_TOKEN_DECIMALS), 100 * (10**ACCEPTED_TOKEN_DECIMALS), specificMaxRaise));
+        vm.expectRevert(
+            abi.encodeWithSelector(
+                ExceedsHardCap.selector,
+                100 * (10 ** ACCEPTED_TOKEN_DECIMALS),
+                100 * (10 ** ACCEPTED_TOKEN_DECIMALS),
+                specificMaxRaise
+            )
+        );
         vm.prank(user2);
         specificRaise.contribute(minAllocToken, new bytes32[](0));
     }
@@ -508,18 +548,20 @@ contract CradleRaiseTest is Test {
 
         // Expect revert with custom error
         // error CancellationWindowPassed(uint256 currentTime, uint256 presaleStart);
-        vm.expectRevert(abi.encodeWithSelector(
-            CancellationWindowPassed.selector,
-            block.timestamp, // Current time after warp
-            raise.presaleStart()
-        ));
+        vm.expectRevert(
+            abi.encodeWithSelector(
+                CancellationWindowPassed.selector,
+                block.timestamp, // Current time after warp
+                raise.presaleStart()
+            )
+        );
 
         // Act
         vm.prank(projectWallet); // Call as owner
         raise.cancelSale();
     }
 
-     function test_CancelSale_PreventsContribution() public {
+    function test_CancelSale_PreventsContribution() public {
         // Arrange
         assertTrue(block.timestamp < raise.presaleStart(), "Timestamp not before presale start");
         vm.prank(projectWallet);
@@ -550,7 +592,7 @@ contract CradleRaiseTest is Test {
         assertTrue(raise.isCancelled(), "Raise should be marked as cancelled");
 
         // Warp time past end time (doesn't really matter as it's cancelled)
-        vm.warp(raise.endTime() + 1 days); 
+        vm.warp(raise.endTime() + 1 days);
 
         // Expect revert because sale is cancelled
         // error SaleNotFinalized(); (Sweep checks !isFinalized || isCancelled)
@@ -571,7 +613,7 @@ contract CradleRaiseTest is Test {
         // Expect event emission
         // event RaiseFinalized(uint256 timestamp);
         vm.expectEmit(true, false, false, true); // Check emitter address only
-        emit CradleRaise.RaiseFinalized(block.timestamp); 
+        emit CradleRaise.RaiseFinalized(block.timestamp);
 
         // Act
         vm.prank(projectWallet); // Call as owner
@@ -582,7 +624,7 @@ contract CradleRaiseTest is Test {
         assertFalse(raise.isCancelled(), "Raise should not be marked as cancelled");
     }
 
-     function test_Fail_FinalizeRaise_NotOwner() public {
+    function test_Fail_FinalizeRaise_NotOwner() public {
         // Arrange: Warp time to after endTime
         vm.warp(raise.endTime() + 1 seconds);
 
@@ -600,11 +642,7 @@ contract CradleRaiseTest is Test {
 
         // Expect revert with custom error
         // error SaleNotEnded(uint256 currentTime, uint256 endTime);
-        vm.expectRevert(abi.encodeWithSelector(
-            SaleNotEnded.selector,
-            block.timestamp, 
-            raise.endTime()
-        ));
+        vm.expectRevert(abi.encodeWithSelector(SaleNotEnded.selector, block.timestamp, raise.endTime()));
 
         // Act
         vm.prank(projectWallet); // Call as owner
@@ -616,9 +654,9 @@ contract CradleRaiseTest is Test {
     function test_Sweep_Success() public {
         // Arrange: Contribute some funds first
         vm.warp(raise.publicSaleStart() + 1 seconds); // Warp to public sale
-        uint256 amountToBuy = minAllocToken; 
-        uint256 requiredAccepted = (amountToBuy * pricePerWholeToken) / (10**TOKEN_DECIMALS);
-        
+        uint256 amountToBuy = minAllocToken;
+        uint256 requiredAccepted = (amountToBuy * pricePerWholeToken) / (10 ** TOKEN_DECIMALS);
+
         vm.startPrank(user1);
         acceptedToken.approve(address(raise), requiredAccepted);
         raise.contribute(amountToBuy, new bytes32[](0));
@@ -654,14 +692,22 @@ contract CradleRaiseTest is Test {
         raise.sweep();
 
         // Assert balances after sweep
-        assertEq(acceptedToken.balanceOf(projectWallet), initialOwnerBalance + expectedProjectAmount, "Owner balance incorrect after sweep");
-        assertEq(acceptedToken.balanceOf(feeRecipient), initialFeeRecipientBalance + expectedFee, "Fee recipient balance incorrect after sweep");
+        assertEq(
+            acceptedToken.balanceOf(projectWallet),
+            initialOwnerBalance + expectedProjectAmount,
+            "Owner balance incorrect after sweep"
+        );
+        assertEq(
+            acceptedToken.balanceOf(feeRecipient),
+            initialFeeRecipientBalance + expectedFee,
+            "Fee recipient balance incorrect after sweep"
+        );
         assertEq(acceptedToken.balanceOf(address(raise)), 0, "Raise contract should be empty after sweep");
     }
 
     function test_Sweep_Success_ZeroContributions() public {
-         // Arrange: No contributions made
-        
+        // Arrange: No contributions made
+
         // Finalize the raise
         vm.warp(raise.endTime() + 1 seconds);
         vm.prank(projectWallet);
@@ -688,7 +734,11 @@ contract CradleRaiseTest is Test {
 
         // Assert balances after sweep (should be unchanged)
         assertEq(acceptedToken.balanceOf(projectWallet), initialOwnerBalance, "Owner balance should be unchanged");
-        assertEq(acceptedToken.balanceOf(feeRecipient), initialFeeRecipientBalance, "Fee recipient balance should be unchanged");
+        assertEq(
+            acceptedToken.balanceOf(feeRecipient),
+            initialFeeRecipientBalance,
+            "Fee recipient balance should be unchanged"
+        );
         assertEq(acceptedToken.balanceOf(address(raise)), 0, "Raise contract should remain empty");
     }
 
@@ -702,7 +752,7 @@ contract CradleRaiseTest is Test {
         vm.expectRevert(abi.encodeWithSelector(Ownable.OwnableUnauthorizedAccount.selector, user1));
 
         // Act: Try to sweep as non-owner
-        vm.prank(user1); 
+        vm.prank(user1);
         raise.sweep();
     }
 
@@ -720,20 +770,20 @@ contract CradleRaiseTest is Test {
         raise.sweep();
     }
 
-     function test_Fail_Sweep_Cancelled() public {
+    function test_Fail_Sweep_Cancelled() public {
         // Arrange: Cancel the sale first
         assertTrue(block.timestamp < raise.presaleStart(), "Timestamp not before presale start");
         vm.prank(projectWallet);
-        raise.cancelSale(); 
+        raise.cancelSale();
 
         assertTrue(raise.isFinalized(), "Raise should be finalized after cancellation");
         assertTrue(raise.isCancelled(), "Raise should be marked as cancelled");
-        
+
         // Warp time after end (doesn't strictly matter)
         vm.warp(raise.endTime() + 1 seconds);
 
         // Expect revert with custom error because isCancelled is true
-        // error SaleNotFinalized(); 
+        // error SaleNotFinalized();
         vm.expectRevert(SaleNotFinalized.selector);
 
         // Act: Try to sweep as owner
@@ -754,7 +804,7 @@ contract CradleRaiseTest is Test {
         uint256 _pricePerToken // Price already scaled by accepted token decimals
     ) internal pure returns (uint256) {
         require(_tokenDecimals > 0 && _tokenDecimals <= 36, "Test Calc: Invalid token decimals");
-        uint256 denominator = 10**_tokenDecimals;
+        uint256 denominator = 10 ** _tokenDecimals;
         if (_tokenAmountToBuy == 0) {
             return 0;
         }
@@ -766,7 +816,7 @@ contract CradleRaiseTest is Test {
         numerator = _tokenAmountToBuy * _pricePerToken;
         uint256 requiredAcceptedToken = numerator / denominator;
         // Ensure calculation result is non-zero if input amount is non-zero
-        require(requiredAcceptedToken > 0, "Test Calc: Calc amount is zero"); 
+        require(requiredAcceptedToken > 0, "Test Calc: Calc amount is zero");
         return requiredAcceptedToken;
     }
-} 
+}
